// src/store/slices/userSlice/type.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export interface UserState {
  emailDelegation: any
  user: any;
  loading: boolean;
  isLoggedIn: boolean;
  error: any;
  token: string;
  sallaConnected: boolean;
  merchantID: string;
  storeName: string;
  storeLink: string;
  storeImage: string;
  createdAt: string;
  newStoreImage: string;
}